Refactor: 최적화 - 폴더 분리 - 채팅방 목록 api 호출 최소화
