+++ Underworld_LINE_Plus_FE/src/pages/main/components/info/room/RoomComponent.jsx
@@ -1,11 +1,6 @@
-import { useEffect } from "react";
 import RoomInfo from "./RoomInfo";
 
-const RoomComponent = ({ userId, setRoomId, roomList, handleRoomList }) => {
-  useEffect(() => {
-    handleRoomList();
-  }, []);
-
+const RoomComponent = ({ userId, setRoomId, roomList }) => {
   return (
     <>
       <RoomInfo userId={userId} roomList={roomList} setRoomId={setRoomId} />
